refactor(control-client): extract useStreamwallWebsocketConnection into its own module
Moves the hook out of index.tsx (which has a top-level render() side
effect on import) into its own file, so it can be imported and tested
in isolation. No behavior change.
